requirements.txt
agate==1.9.1
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiohttp-cors==0.8.1
aiosignal==1.3.2
aiosqlite==0.20.0
alembic==1.14.1
annotated-types==0.7.0
antlr4-python3-runtime==4.11.1
anyio==4.8.0
apache-airflow==2.10.5
apache-airflow-providers-common-compat==1.7.1
apache-airflow-providers-common-io==1.5.0
apache-airflow-providers-common-sql==1.27.1
apache-airflow-providers-fab==1.5.2
apache-airflow-providers-ftp==3.12.0
apache-airflow-providers-google==16.0.0
apache-airflow-providers-http==5.0.0
apache-airflow-providers-imap==3.8.0
apache-airflow-providers-postgres==6.2.0
apache-airflow-providers-slack==9.1.2
apache-airflow-providers-smtp==1.9.0
apache-airflow-providers-sqlite==4.0.0
apispec==6.8.1
argcomplete==3.5.3
asgiref==3.8.1
async-timeout==5.0.1
asyncpg==0.30.0
attrs==25.1.0
babel==2.17.0
backoff==2.2.1
blinker==1.9.0
cachelib==0.9.0
cachetools==5.5.2
cattrs==25.1.1
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
clickclick==20.10.2
colorama==0.4.6
colorful==0.5.6
colorlog==6.9.0
ConfigUpdater==3.2
connexion==2.14.2
cron-descriptor==1.4.5
croniter==6.0.0
cryptography==42.0.8
daff==1.4.2
db-dtypes==1.4.3
dbt-adapters==1.16.1
dbt-bigquery==1.10.0
dbt-common==1.26.0
dbt-core==1.10.4
dbt-extractor==0.6.0
dbt-protos==1.0.337
dbt-semantic-interfaces==0.9.0
decorator==5.2.1
deepdiff==7.0.1
Deprecated==1.2.18
dill==0.3.1.1
distlib==0.3.9
dnspython==2.7.0
docstring_parser==0.16
email_validator==2.2.0
exceptiongroup==1.2.2
fastjsonschema==2.21.1
filelock==3.18.0
Flask==2.2.5
Flask-AppBuilder==4.5.2
Flask-Babel==2.0.0
Flask-Caching==2.3.0
Flask-JWT-Extended==4.7.1
Flask-Limiter==3.10.1
Flask-Login==0.6.3
Flask-Session==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.2.2
frozenlist==1.5.0
fsspec==2025.5.1
gcloud-aio-auth==5.4.2
gcloud-aio-bigquery==7.1.0
gcloud-aio-storage==9.4.0
gcsfs==2025.5.1
google-ads==27.0.0
google-analytics-admin==0.24.0
google-api-core==2.25.1
google-api-python-client==2.173.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-cloud-aiplatform==1.98.0
google-cloud-alloydb==0.4.7
google-cloud-appengine-logging==1.6.2
google-cloud-audit-log==0.3.2
google-cloud-automl==2.16.4
google-cloud-batch==0.17.36
google-cloud-bigquery==3.34.0
google-cloud-bigquery-datatransfer==3.19.2
google-cloud-bigquery-storage==2.32.0
google-cloud-bigtable==2.31.0
google-cloud-build==3.31.2
google-cloud-compute==1.31.0
google-cloud-container==2.57.0
google-cloud-core==2.4.3
google-cloud-datacatalog==3.27.1
google-cloud-dataflow-client==0.9.0
google-cloud-dataform==0.6.2
google-cloud-dataplex==2.10.2
google-cloud-dataproc==5.20.0
google-cloud-dataproc-metastore==1.18.3
google-cloud-dlp==3.30.0
google-cloud-kms==3.5.1
google-cloud-language==2.17.2
google-cloud-logging==3.12.1
google-cloud-managedkafka==0.1.11
google-cloud-memcache==1.12.2
google-cloud-monitoring==2.27.2
google-cloud-orchestration-airflow==1.17.5
google-cloud-os-login==2.17.2
google-cloud-pubsub==2.30.0
google-cloud-redis==2.18.1
google-cloud-resource-manager==1.14.2
google-cloud-run==0.10.18
google-cloud-secret-manager==2.24.0
google-cloud-spanner==3.55.0
google-cloud-speech==2.33.0
google-cloud-storage==2.19.0
google-cloud-storage-transfer==1.17.0
google-cloud-tasks==2.19.3
google-cloud-texttospeech==2.27.0
google-cloud-translate==3.20.3
google-cloud-videointelligence==2.16.2
google-cloud-vision==3.10.2
google-cloud-workflows==1.18.2
google-crc32c==1.7.1
google-genai==1.21.1
google-re2==1.1.20240702
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
grpc-google-iam-v1==0.14.2
grpc-interceptor==0.15.4
grpcio==1.73.0
grpcio-gcp==0.2.2
grpcio-status==1.71.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
immutabledict==4.2.1
importlib-metadata==6.11.0
inflect==7.5.0
inflection==0.5.1
isodate==0.6.1
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==0.10.0
joblib==1.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_core==5.8.1
lazy-object-proxy==1.10.0
leather==0.4.0
limits==4.0.1
linkify-it-py==2.0.3
lockfile==0.12.2
looker-sdk==25.10.0
Mako==1.3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.0
marshmallow-oneofschema==3.1.1
marshmallow-sqlalchemy==0.28.2
mashumaro==3.14
mdit-py-plugins==0.4.2
mdurl==0.1.2
methodtools==0.4.7
more-itertools==10.6.0
msgpack==1.1.1
multidict==6.1.0
nbformat==5.10.4
networkx==3.4.2
numpy==2.2.6
oauthlib==3.3.1
opencensus==0.11.4
opencensus-context==0.1.3
opentelemetry-api==1.34.1
opentelemetry-exporter-otlp==1.34.1
opentelemetry-exporter-otlp-proto-common==1.34.1
opentelemetry-exporter-otlp-proto-grpc==1.34.1
opentelemetry-exporter-otlp-proto-http==1.34.1
opentelemetry-proto==1.34.1
opentelemetry-sdk==1.34.1
opentelemetry-semantic-conventions==0.55b1
ordered-set==4.1.0
packaging==24.2
pandas==2.2.3
pandas-gbq==0.29.1
parsedatetime==2.6
pathspec==0.12.1
pendulum==3.0.0
platformdirs==4.3.8
pluggy==1.5.0
prison==0.2.1
prometheus_client==0.22.1
propcache==0.2.1
proto-plus==1.26.1
protobuf==5.29.5
psutil==6.1.1
psycopg2-binary==2.9.10
py-spy==0.4.0
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydata-google-auth==1.9.1
Pygments==2.19.1
PyJWT==2.10.1
pyOpenSSL==25.1.0
pyparsing==3.2.3
python-daemon==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-nvd3==0.16.0
python-slugify==8.0.4
pytimeparse==1.1.8
pytz==2025.1
PyYAML==6.0.2
ray==2.42.0
referencing==0.36.2
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rich==13.9.4
rich-argparse==1.6.0
rpds-py==0.22.3
rsa==4.9.1
ruamel.yaml==0.17.40
ruamel.yaml.clib==0.2.12
scikit-learn==1.5.2
scipy==1.15.3
setproctitle==1.3.4
shapely==2.1.1
six==1.17.0
slack_sdk==3.36.0
smart-open==7.1.0
sniffio==1.3.1
snowplow-tracker==1.1.0
soda-core==3.5.5
soda-core-bigquery==3.5.5
soda-core-postgres==3.5.5
SQLAlchemy==1.4.54
sqlalchemy-bigquery==1.14.1
SQLAlchemy-JSONField==1.0.2
sqlalchemy-spanner==1.13.1
SQLAlchemy-Utils==0.41.2
sqlparse==0.5.3
tabulate==0.9.0
tenacity==8.5.0
termcolor==2.5.0
text-unidecode==1.3
threadpoolctl==3.6.0
time-machine==2.16.0
tqdm==4.67.1
traitlets==5.14.3
typeguard==4.4.3
types-protobuf==6.30.2.20250516
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.1
uc-micro-py==1.0.3
universal_pathlib==0.2.6
uritemplate==4.2.0
urllib3==2.3.0
virtualenv==20.31.2
websockets==15.0.1
Werkzeug==2.2.3
wirerope==1.0.0
wrapt==1.17.2
WTForms==3.2.1
yarl==1.18.3
zipp==3.21.0
